Feature document bulk

See merge request dms/backend!7

+++ app/Services/Document/DocumentService.php
@@ -5,10 +5,13 @@ namespace App\Services\Document;
 use App\Contracts\Repositories\IDocumentRepository;
 use App\Document;
 use App\DocumentVersion;
-use App\Http\Controllers\API\DocumentVersionController;
+use App\Events\Document\DocumentCreateEvent;
+use App\Events\Document\DocumentDeleteEvent;
+use App\Events\Document\DocumentReadEvent;
+use App\Events\Document\DocumentUpdateEvent;
+use App\Services\Components\IEventDispatcher;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 /**
  * @author Vladimir Barmotin <barmotinvladimir@gmail.com>
@@ -20,23 +23,35 @@ class DocumentService
      * @var DocumentVersionService
      */
     private $documentVersionService;
+    /**
+     * @var IEventDispatcher
+     */
+    private $eventDispatcher;
 
     /**
      * DocumentService constructor.
      *
      * @param IDocumentRepository $repository
      * @param DocumentVersionService $documentVersionService
+     * @param IEventDispatcher $eventDispatcher
      */
-    public function __construct(IDocumentRepository $repository, DocumentVersionService $documentVersionService)
+    public function __construct(
+        IDocumentRepository $repository,
+        DocumentVersionService $documentVersionService,
+        IEventDispatcher $eventDispatcher
+    )
     {
         $this->repository = $repository;
         $this->documentVersionService = $documentVersionService;
+        $this->eventDispatcher = $eventDispatcher;
     }
 
     public function create(Document $document): Document
     {
         $id = $this->doCreate($document);
         $document->setId($id);
+
+        $this->eventDispatcher->dispatch(new DocumentCreateEvent($document));
 
         return $document;
     }
@@ -63,7 +78,10 @@ class DocumentService
 
     public function get($id): Document
     {
-        return $this->repository->findOrFail($id);
+        $document = $this->repository->findOrFail($id);
+        $this->eventDispatcher->dispatch(new DocumentReadEvent($document));
+
+        return $document;
     }
 
     public function update(int $id, Document $documentInput, $updatedFields, bool $createNewVersion = true): Document
@@ -74,40 +92,55 @@ class DocumentService
             $document->{dms_build_setter($fieldKey)}($documentInput->{dms_build_getter($fieldKey)}());
         }
 
-
-        $oldActualVersion = $this->repository->getActualVersionRelation($document);
         $newActualVersion = $documentInput->getActualVersion();
-        $newActualVersion->setDocumentId($id);
-        $newActualVersion->setActual(true);
 
-        $this->doUpdate($document, $oldActualVersion, $newActualVersion, $createNewVersion);
+        if($newActualVersion) {
+            $oldActualVersion = $this->repository->getActualVersionRelation($document);
+            $newActualVersion->setDocumentId($id);
+            $newActualVersion->setActual(true);
 
+            $this->doUpdate($document, $oldActualVersion, $newActualVersion, $createNewVersion);
+        } else {
+            $this->doUpdate($document, null, null, false);
+        }
+
+        $this->eventDispatcher->dispatch(new DocumentUpdateEvent($document));
         return $this->get($id);
     }
 
-    protected function doUpdate(Document $document, DocumentVersion $oldActualVersion, DocumentVersion $newActualVersion, bool $createNewVersion): void
+    protected function doUpdate(Document $document, ?DocumentVersion $oldActualVersion, ?DocumentVersion $newActualVersion, bool $createNewVersion): void
     {
         $this->repository->save($document);
-        if($createNewVersion) {
-            $oldActualVersion->setActual(false);
-            $this->documentVersionService->update($oldActualVersion);
-            $newActualVersion->setVersionName((int)$oldActualVersion->getVersionName() + 1);
-        } else {
-            $this->documentVersionService->delete($oldActualVersion->getId(), true);
-            $newActualVersion->setVersionName($oldActualVersion->getVersionName());
+
+        if($newActualVersion && $oldActualVersion) {
+            if ($createNewVersion) {
+                $oldActualVersion->setActual(false);
+                $this->documentVersionService->update($oldActualVersion);
+                $newActualVersion->setVersionName((int)$oldActualVersion->getVersionName() + 1);
+            } else {
+                $this->documentVersionService->delete($oldActualVersion->getId(), true);
+                $newActualVersion->setVersionName($oldActualVersion->getVersionName());
+            }
+            $this->documentVersionService->create($newActualVersion);
         }
 
-        $this->documentVersionService->create($newActualVersion);
     }
 
     public function delete(int $id): ?bool
     {
-        return $this->repository->delete($this->get($id));
+        $document = $this->repository->find($id);
+
+        if(!$document) {
+            return false;
+        }
+
+        $this->eventDispatcher->dispatch(new DocumentDeleteEvent($document));
+        return $this->repository->delete($document);
     }
 
-    public function list(): LengthAwarePaginator
+    public function list(array $filters): LengthAwarePaginator
     {
-        return $this->repository->list();
+        return $this->repository->list($filters);
     }
 
     public function setActualVersion($documentId, $documentVersionId)
